+++ src/app/apply/page.tsx
@@ -1,3 +1,13 @@
+import WriteStep from "@/components/domains/apply/steps/WriteStep";
+import RoleStep from "@/components/domains/apply/steps/RoleStep";
+import SuccessStep from "@/components/domains/apply/steps/SucessStep";
+
 export default function ApplyPage() {
-  return null;
+  return (
+    <>
+      <WriteStep />
+      {/* <RoleStep /> */}
+      {/* <SuccessStep /> */}
+    </>
+  );
 }
